# Conflicts:
#	Group3.Semester3.WebApp/BusinessLayer/FileService.cs
#	Group3.Semester3.WebApp/Controllers/Api/FileApiController.cs

+++ Group3.Semester3.WebApp/BusinessLayer/FileService.cs
@@ -1,5 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IO;
+using System.Text;
 using System.Threading.Tasks;
 using Group3.Semester3.WebApp.Helpers;
 using Group3.Semester3.WebApp.Models.Users;
@@ -10,6 +12,9 @@ using Group3.Semester3.WebApp.Entities;
 using Group3.Semester3.WebApp.Repositories;
 using Microsoft.AspNetCore.Http;
 using Group3.Semester3.WebApp.Helpers.Exceptions;
+using Azure.Storage.Sas;
+using Azure.Storage;
+using Azure.Storage.Blobs.Specialized;
 
 namespace Group3.Semester3.WebApp.BusinessLayer
 {
@@ -22,7 +27,11 @@ namespace Group3.Semester3.WebApp.BusinessLayer
         public bool DeleteFile(Guid fileId, UserModel user);
         public FileEntity CreateFolder(UserModel user, CreateFolderModel model);
         public bool MoveIntoFolder(FileEntity model, UserModel user);
+        public (FileEntity, string) DownloadFile(Guid fileId, Guid userId);
+        public UpdateFileModel GetFileContents(string id, UserModel user);
+        public FileEntity UpdateFileContents(UpdateFileModel model, UserModel user);
     }
+    
     public class FileService : IFileService
     {
         private IConfiguration _configuration;
@@ -70,25 +79,26 @@ namespace Group3.Semester3.WebApp.BusinessLayer
             {
                 if (formFile.Length > 0)
                 {
-                    var blobName = Guid.NewGuid();
+                    var blobGuid = Guid.NewGuid();
                     try
                     {
-                        await containerClient.UploadBlobAsync(blobName.ToString(), formFile.OpenReadStream());
+                        await containerClient.UploadBlobAsync(blobGuid.ToString(), formFile.OpenReadStream());
                         fileEntries.Add(new FileEntry
                         {
                             Name = formFile.FileName,
-                            Id = blobName,
+                            Id = blobGuid,
                             Parent = new DirectoryEntry { Id = parsedGUID }
                         });
 
                         var file = new FileEntity()
                         {
                             Id = Guid.NewGuid(),
-                            AzureId = blobName,
+                            AzureName = blobGuid.ToString(),
                             Name = formFile.FileName,
                             UserId = user.Id,
                             ParentId = parsedGUID,
-                            IsFolder = false
+                            IsFolder = false,
+                            Updated = DateTime.Now
                         };
 
                         _fileRepository.Insert(file);
@@ -106,16 +116,68 @@ namespace Group3.Semester3.WebApp.BusinessLayer
             return fileEntries;
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="fileId"></param>
+        /// <param name="userId"></param>
+        /// <returns>A tuple with the FileEntity with the given fileId and a string with the download URL</returns>
+        public (FileEntity, string) DownloadFile(Guid fileId, Guid userId)
+        {
+            var file = _fileRepository.GetById(fileId);
+
+            if (userId == file.UserId)
+            {
+                BlobContainerClient containerClient =
+                new BlobContainerClient(
+                    _configuration.GetConnectionString("AzureConnectionString"),
+                    _configuration.GetSection("AppSettings").Get<AppSettings>().AzureDefaultContainer);
+
+                containerClient.CreateIfNotExists();
+
+                BlobSasBuilder blobSasBuilder = new BlobSasBuilder()
+                {
+                    StartsOn = DateTime.UtcNow,
+                    ExpiresOn = DateTime.UtcNow.AddHours(24),
+                    BlobContainerName = containerClient.Name,
+                    BlobName = file.AzureName,
+                    Resource = "b"
+                };
+
+                blobSasBuilder.SetPermissions(BlobContainerSasPermissions.Read);
+
+                StorageSharedKeyCredential storageSharedKeyCredential = new StorageSharedKeyCredential(
+                    _configuration.GetSection("AppSettings").Get<AppSettings>().AzureStorageAccount,
+                    _configuration.GetSection("AppSettings").Get<AppSettings>().AzureAccountKey);
+
+                string sasToken = blobSasBuilder.ToSasQueryParameters(storageSharedKeyCredential).ToString();
+
+                return (file, $"{containerClient.GetBlockBlobClient(file.AzureName).Uri}?{sasToken}");
+            }
+            else throw new ValidationException("Operation forbidden.");
+        }
+
         public bool DeleteFile(Guid fileId, UserModel user)
         {
-            var file = GetById(fileId);
-            _accessService.hasAccess(user, file);
-            var result = _fileRepository.Delete(fileId);
-            if(!result)
+            BlobContainerClient containerClient =
+                new BlobContainerClient(
+                    _configuration.GetConnectionString("AzureConnectionString"),
+                    _configuration.GetSection("AppSettings").Get<AppSettings>().AzureDefaultContainer);
+
+            var file = _fileRepository.GetById(fileId);
+            if (userId == file.UserId)
             {
-                throw new ValidationException("File non-existent or not deleted.");
+                // TODO AzureId should be string
+                containerClient.DeleteBlob(_fileRepository.GetById(fileId).AzureName.ToString());
+                var result = _fileRepository.Delete(fileId);
+
+                if (!result)
+                {
+                    throw new ValidationException("File non-existent or not deleted.");
+                }
+                else return result;
             }
-            else return result;
+            else throw new ValidationException("Operation forbidden.");
         }
 
         public FileEntity RenameFile(Guid fileId, UserModel user, string name)
@@ -169,7 +231,7 @@ namespace Group3.Semester3.WebApp.BusinessLayer
             {
                 Id = Guid.NewGuid(),
                 Name = model.Name,
-                AzureId = Guid.Empty,
+                AzureName = string.Empty,
                 UserId = user.Id,
                 ParentId = parentGuid,
                 IsFolder = true
@@ -195,6 +257,77 @@ namespace Group3.Semester3.WebApp.BusinessLayer
                 throw new ValidationException("File has not been moved, try again.");
             }
             else return true;
+        }
+
+        public UpdateFileModel GetFileContents(string id, UserModel user)
+        {
+            var fileId = ParseGuid(id);
+            
+            var file = _fileRepository.GetById(fileId);
+            
+            if (file.UserId != user.Id)
+            {
+                throw new ValidationException("Unauthorized");
+            }
+            
+            var containerClient =
+                new BlobContainerClient(
+                    _configuration.GetConnectionString("AzureConnectionString"),
+                    _configuration.GetSection("AppSettings").Get<AppSettings>().AzureDefaultContainer);
+
+            containerClient.CreateIfNotExists();
+
+            var response = containerClient.GetBlobClient(file.AzureName).Download();
+            var stream = response.Value.Content;
+
+            StreamReader reader = new StreamReader(stream);
+            string text = reader.ReadToEnd();
+
+            var model = new UpdateFileModel()
+            {
+                Id = file.Id,
+                Contents = text,
+                Timestamp = DateTime.Now
+            };
+
+            return model;
+        }
+
+        public FileEntity UpdateFileContents(UpdateFileModel model, UserModel user)
+        {
+            var file = _fileRepository.GetById(model.Id);
+
+            if (!model.Overwrite)
+            {
+                var result = DateTime.Compare(model.Timestamp, file.Updated);
+
+                if (result <= 0)
+                {
+                    throw new ConcurrencyException("File was changed by another user. Please try again");
+                }
+            }
+
+            if (file.UserId != user.Id)
+            {
+                throw new ValidationException("Unauthorized");
+            }
+            
+            byte[] byteArray = Encoding.ASCII.GetBytes( model.Contents );
+            var contentStream = new MemoryStream( byteArray );
+
+            var containerClient =
+                new BlobContainerClient(
+                    _configuration.GetConnectionString("AzureConnectionString"),
+                    _configuration.GetSection("AppSettings").Get<AppSettings>().AzureDefaultContainer);
+
+            containerClient.CreateIfNotExists();
+
+            // Trigger update change
+            _fileRepository.Rename(file.Id, file.Name);
+
+            containerClient.GetBlobClient(file.AzureName).Upload(contentStream, true);
+
+            return file;
         }
     }
 }
